Log APK hash for package install attempts in the Event Log.
This CL adds a package_digest field to the install_package_attempt
event. The field is populated with the SHA-256 digest of the contents
of the APK iff the user has consented to app verification and app
verification is enabled.

Bug: 11275004
Bug: 10605940

(cherry picked from commit 8fca480bee00578c1529b1f32ab30af096db82dd)

Change-Id: I9773925f7e397ada26efac022349dc8e4af01208