Support uninstalling apps for other profiles.
Allow client to pick which user the package
will be removed for, checking its a profile
of the current user.
Inform user if package is blocked from being
removed and let them knows its due to an
admin.
Add check to stop non owner asking to uninstall
for all users.

Bug: 14127299
Change-Id: I60504224f5271272a390320a0fa62aa7f5de4e54