Do not disable app ops on a permission revocation for modern apps.
We were disabling both permission and app ops for modern apps - to
have app ops as a safety net. This creates the issue that if a
permission is revoked (which disables the app op) and data is
cleared, then apps that get permissions granted by default would
end up with th permission but not the app op and will not work.
We still enable the app op on a permission grant to ensure that
a legacy app with a disabled app ops will have the app op enabled
if updated and the permission is then granted.

bug:22464566

Change-Id: Ia5f4de42bcaf6f736e50d6f3109878ff2ce8ff9f