Build initial Permissions Settings UI page for Wear
This is a first pass to build a functional settings UI page
for Permissions on Android Wear.  The user flow is a little
different than the phone version, so there are some nuances
that need to be adjusted from the normal flow.

I have forked off of the AppPermissionsFragment and created
a Wear version for us to use.  Rather than try to link
the Wearable-Support lib into AOSP code, I have made a
duplicate of the few files I needed.

This change adds a Wear layout version of the Settings UI
page.  The UI is fully functional for most cases, but there
are some tweaks still needed for full parity.  The tweaks
are primarily around UI and special edge case functions.

BUG: 23080561
Change-Id: I8477f6b966cacaae9e77aa0fb61b4b1e621a9ead