Clean up package installer styling
Package installer should be using exclusively the device
defaut themes on handhelds (phones and tables) to allow
OEMs customize its look and feel. Also for handhelds the
installer should use the same preference framework as the
settings app to ensure it looks like the settings app on
an OEMs device.

This change needs to be picked up by a GMS build to deliver
to partners ensuring the installer UI is consistent with
the device UI.

bug:24286616

Change-Id: I92e39fd1488e76b0b23b7f1efa13e04ed5bbc7ba