Make the permission request dialog's layout robust
The old implementation was relying on a fixed window size
where the content is positioned by a custom layout manager.
It is possible however that subsequent permissions requests
do not fit in the window as its size is computed based on
the content of the first permissions request. There were
also cases where the content is chopped after a rotation
as the dialog size width was not re-evaluated while it should
be. Further, animation from one permission request state
to another was not properly done resulting in content
being chopped off in some cases.

The current approach is to have a dialog width for the
content activity but the height is as tall as the screen
allowing us to fit arbitrary large permission request
content. Also we are resetting the fixed width on a
configuration change so the dialog is robust to adjust
size as needed.

bug:24679384
bug:25755378

Change-Id: I9e885f55f2234f4156fb62e92c7e3d3e6e268ae8