Introduce GrantPermissionViewHandler for Auto usecase.
For Android Auto Embedded, the permissions dialog needs to have larger
font and tap targets for the buttons. In order to accomplish this, the
dialog must be allowed to expand to at most the full width of the
screen.

Allow GrantPermissionsViewHandlerImpl to be extended to reduce the
amount of code that needs to be duplicated.

Refer to ag/902867 for the AAE's overlay that replaces the
grant_permissions.xml and permissions_description.xml layouts.

Bug: 27967879
Change-Id: Ie30fcc5d1e7fa9cb22da8faf40c70daccf840f6f