Fix app install flow
A regression was introduced when we added support for
installing apps from content URIs. Since we stage the
APK in the cache folder we changed the flow to wipe
the cached version when the PackageInstallerActivity
finishes or gets an installation result, so we started
the InstallAppProgress activity for result but the latter
may be called with the forward result flag which is
incompatible with start activity for result. This change
delegates clearing the staging file in the cache folder
to the InstallAppProgress activity.

bug:28551937

Change-Id: I0d9978aff60b7fab6b64fe7bf889ef30f9b2fd7c