[DO NOT MERGE] Bring back overlay warning in permissions UI.
The package installer has a special event handling when
the user tries to interact with it and an overlay window
intersects the installer one. We show a warning that an
overlay was detected and point the user to settings. This
regressed and nothing happened when the users click the
allow button.

bug:28572707
bug:26677796

Change-Id: I617595e7cd5277576ccb9915946b77082a1c2b9e

Resolution to mnc-mr2-release when CP'ing ag/1065580

Change-Id: Ia4d603a88bb4d7c38ec9ecadad1f4eac9e697516