Remove lefover secure button
An incorrect fix for a security bug where apps with
overlay access can redress the permission message added
a secure button implementation that refuses interaction
if the window is covered. While this was fixed in M
branches we did not merge the change in NYC because
we now hide all overlays. However, the app op which
controls overlays allows privileged apps to bypass the
app op. It is fine to have priv apps overlay stuff on
top of the permissions UI except that the allow button
show be clickable. This change removes the offending
code.

bug:29879858

Change-Id: I9d0b5e62b5ef029069e37f34dbcd90d4909d966b