Reworked additional permissions fragment for each app
Fixed packageinstaller crash when selecting additional permissions for
an app e.g. "Live Channels". setHeader should not be called in onCreate
since it will update title UI which is initialized in onCreateView in
SettingsWithHeader.java.
Also, updated the UI for AdditionalPermissionsFragment for each
individual app to include the app banner and title in the heading.

Change-Id: Idfc12a28b38b84fcc5d97de4f85ada968e122d40
Fixes: 32338150
Test: Visual inspection of the permission UI