Properly set permission flags in review mode for modern apps
We were not properly setting the permission flags in review mode
if the user manages individual permissions in a group. Specifically,
if the user revokes an individual permission in the group it has
to be marked as fixed so a subsequent request would not grant it
because there is a permission in the group already granted. Also
if the last permissions in the group is individually revoked we
need to clear the fixed flag for all permissions in the group.
Further, if a permission is granted in the group then similarly
we need to mark the rest in the group fixed to prevent an auto
grant. Lastly, if all permissions in a group are granted we need
to clear the fixed flag for all of them.

Test: manually flipped all combinations and dumped the flags

Change-Id: Ic5d8f88aa33c8cb029a3bd19955ae3184fcc4a9b