Fix missing case to update permission group subtitle
Permissions in some groups are individually controlled if
permission review mode is enabled. We were not properly
updating the subtitle of a group with individually controlled
permissions.

Also while at this fix a bug where we grnated an intividually
controller group if the user granted two permissions in the
group.

Test: manual

bug:33463159
bug:35622170

Change-Id: I0913531d69668ccf2c2fdd3cdf0d9e9a42e67dc0