Allow permission grant UI to show on top of keyguard
Interestingly, setting showForAllUsers which is supposed to be
a repacement for showOnLockScreen does not work. Further, the
latter seems to be not only deprecated but also completely
ignored, potentially breaking apps. Anyway, setting the flags
(FLAG_SHOW_WHEN_LOCKED, FLAG_TURN_SCREEN_ON) works fine. This
change allows apps that show on top of keyguard to ask for
permissions - a use case is a communication app receiving a
call but needing a permission to handle it.

Test: manual

bug:38312872

Change-Id: I22457ee8b0e68ebd81e7e5f0f0aa216d8f83b9d1