DO NOT MERGE Fixing Install button not enabled on resume
In cases where an app had a long list of permissions the Install button
was not being enabled.

Test: Manually clicked on an apk that required a long list of
permissions. Checked that the install button is enabled. Switch out and
back to the activity to force a pause/resume. The button gets disabled
and enabled as expected. Also the existing test passes as before:
gts-tradefed run gts-dev -m PackageInstallerTapjacking

Bug: 35056974
Change-Id: I2ef7d5ce4e137969c70d2929271620c2bcb723f2