Do not start staging task if activity is finishing
If an app is resumed after the process was killed onResume is called
even if onActivityResult sets finish() before. This does not happen if
the activity is resumed without the process having been killed.

Obvisouly we do not want to start an installation if the activity is
finishing, hence just don't do anything in this case.

Fixes: 67759288
Test: - Went through installation until final screen
      - Put installer into background
      - Killed package installer manually
      - Recents -> back to installer
      - Saw onResume called but installer finished anyway

Change-Id: Ib1347692de9deb20459b71410b0ab21cc9e11e06