Allow to grant only foreground permission
If both background and foreground permissions are requested we force the
user to make a choice by showing a radio-group with the choices
always/foreground/deny. No option is preselected.

If only foreground or background is requested we just show an additional
message explaining the foreground/background behavior.

Internally to GrantPermissionsActivity foreground groups and background
groups are handled separately, i.e. they have a separate GroupState as
one might be granted, one might be denied. The only time they get merged
is when we prepare the UI as we only want to show a single UI for
background+foreground.

Test: atest PermissionsHostTest
      + manual testing
Bug: 78788390
Change-Id: Ieb617d73a3d1247c05207872d1c1980fd33d754c