Fixes two bugs in background permission handling
(1) In AppPermissionGroup the isAppOpAllowed property of background
permissions is based on the app op state of the foregrond permission.
Hence it needs to be set once we found the link between the background
and foreground permissions.

(2) In PermissionPreference we might show a deny-default dialog is the
app does not support runtime permissions. Hence we also have to notify
the that the deny-default was approved in the same case.

Test: atest PermissionsHostTests
Change-Id: I4cd386dc557ab6fffbdc77b0f880d4c1c03341dd