Add the permission usage screen of the Permissions Hub.
This adds an initial version of the permission usage screen.  It shows
permission requests by apps with the ability to filter based on
permission, time, and whether or not to show system apps.

Currently the only way to access it is through searching Settings for
"permissions usage", and clicking on an entry brings you to that app's
list of permissions.

Bug: 63532550
Test: Used the Permissions Hub.

Change-Id: I9282f859cb99d2ade4cb9b01d8145ca27e17363e