Warn every time when revoking default granted permissions from list of apps
(1)
When looking at a list of apps and revoking permissions it is not clear
which permissions are granted by default and which are not. Hence we
should warn before a user revoked permissions that are granted by
default every time.

(2)
This is different from looking at an app. In this case permissions are
usually all granted by default or not granted. Hence in this case we
only warn once and not when the user revoked the second permission.

We used to only show the warning once in the first case, but should not.
This changes fixes that.

Test: Revoked granted-by-default permissions from list of all apps and
from list of all permissions of a app.

Change-Id: Ie0481a8c3936b82934b534ff172ff03d10e715c5