Have AppPermissionFragment listen for changes to its package.
Modify AppPermissionFragment to listen for permission changes and
package uninstalls.  If a permission changes, it changes the selected
radio button.  It the package is uninstalled, the activity exits.

This also fixes PackageRemovalMonitor so that it properly receives
uninstall events.

Bug: 63532550
Test: Open page, adb revoke permission, see selected button change
Test: Open page, adb uninstall app, see activity exit Test: Open page,
return to home, adb uninstall app, switch back to page, see activity
exit

Change-Id: Ia9f080aaa323910c1c35ed648f8d6c8bbb1c6603