Prevent spinner overlaps with large font size.
This stops the two spinners in PermissionUsageFragment from
overlapping when the font or display sizes are set to large.  Using
ButtonBarLayout allows them to switch to being stacked vertically when
there is not enough space for them horizontally.

This slightly modified the behavior of ButtonBarLayout to prevent them
from being right-aligned when stacked vertically.

Bug: 119771291
Test: Set font and display size to large and see no overlap.
Change-Id: I8ae9bc9cdb49d728aaa09d5238ba31df36c62997