Have most uses of PermissionPreference go to AppPermissionFragment.
Part of the Permissions Hub redesign modifies the existing
AppPermissionsFragment and PermissionsAppFragment screens to contain
links to the new AppPermissionFragment instead of the inlined toggles
of PermissionPreference.  This implements that change.

In addition, I added the subtitle for individually-controlled
preferences to AppPermissionFragment and updated to the latest mocks,
which show the most recent usage in the subtitle in
AppPermissionsFragment.

Bug: 63532550
Test: Open both fragments, change a permission, verify that the
preference is moved to the proper segment.

Change-Id: I7001e76532426f58581ba7000846117d07cff043