Fix Permissions Hub handling of work profile.
If users had the same app installed in the normal and the work
profiles, PermissionUsageFragment previously only showed one of them
(whichever was accessed more recently).  I modified it to
differentiate between different uids.

In addition, AppPermissionFragment only worked on non-work profile
apps.  I modified it so that it can take in the users and uses that to
get the correct app.

Bug: 119811056
Test: Setup work profile, use both Contacts apps, give them different
permissions, then see separate entries in PermissionUsageFragment and
verify that AppPermissionFragment can change either app's permissions.

Change-Id: I448c26850069ce539cb1afa82f72807da779da84