Revert "Revert "Add work profile support for the new Default apps UI.""
This reverts commit 1da06cd0b762033234d1f5966b0672de327650d2.

Reason for revert: The change itself doesn't break any build other than mainline_modules-userdebug, and that build breakage is likely caused by building this change against an old SDK which didn't have the new APIs. Once we make sure we are building with a new SDK we can merge this double-revert.

Change-Id: I07c7231ec9fb0f0c2eb56b590a513e16ba048916