Update PermissionController to use the new historical ops API
The historical ops API changed and we had to update the code
to use the new APIs. The new APIs now read history from the
disk which would be a problem with the current impl where we
make multiple calls into the app ops manager - one per app.
It is more efficient to make a single call for all data we
care about. This change adds a new mechanism to load the data
once and off the main thread as the IPC would hit the disk.

Test: manual

bug:111061782

Change-Id: I86113929fdd4a80fe620ed3eccd591712de30f47