Run LocationAccessCheck as parent user
If the current user has a work profile we only want to show one
location-access-check notification at a time, not one for the personal
and one for the work profile.

Hence we should run the location-access-check code always as the parent
user. The code already handled the other profiles by creating alternate
manager-s for the other users if needed.

Test: atest --test-mapping packages/apps/PermissionController/src/com/android/packageinstaller/permission/service:presubmit
Change-Id: I6339d46bd402878ce9da597b7c81f61b521758f3