Update to a newer version of the Permission Hub mocks.
This gives singleton/uncoalesced entries "parent" permissions just
like the coalesced ones to be more consistent.  It also changes the
summary, title, and icon of the entries themselves, including changing
to absolute time for entries in the current day and the date for older
entries.

It also changes the default sort to recency, adds a one minute filter,
and continues to remove the Storage permission.

Bug: 63532550
Test: Open Permissions Hub
Change-Id: Ib37a93d0bc47eaa6a74d5df192e833cfb13949e1