Prevent creating AppPermissionGroup for wrong user
AppPermissionGroup#create takes a PackageInfo that contains the
permissions for an app (which is per user). It also take a
userHandle argument. Hence it can happen that the userHandle and
the permissions do not match.

Hence remove all userHandle arguments from the creation path to avoid
this problem in the first place.

Test: Granted and revoked permission in a multi-user environment. Checked
      permission screens.

Change-Id: I53307538f91bf7c95b571dadf3a9d098193800c6