Use full group label in AppPermissionsFragment.
When AppPermissionGroup loaded the label of its group, it was calling
loadLabel, which was truncating the string.  Let's also have it load
the full label, and use that in AppPermissionsFragment to avoid
truncating the group name.

Fixes: 123484592
Test: Change language to French, open the list of permission for Maps,
see full name of "Activity recognition" group (but in French).
Test: Set font size and display size to large and see group labels
wrap to next line in French.

Change-Id: I9fdb998e0664b6ad01e5fe65d4dff023e367e08d