Fix browser role fallback.
Previously we always return the factory default browser as a
fallback (if there is one), however in most cases this will break the
previous behavior of clearing the default browser when a new browser
is installed. This change changes this behavior to be setting a
fallback only if there is only one browser.

Test: manual
Change-Id: Iac142c2ec0a4fb6eecb55576af817c9589558fe0