Optimize Permissions Hub and App Permissions screens.
The majority of the time spent loading the Permissions Hub was spent loading the labels and icons for all apps.  However, the screen does not show most apps, as they either have no uses or are system (which are not shown by default).  This commit makes us only load labels and icons for apps we will show.  We do the same thing for App Permissions, as it has the same issue.

On my dev device, loading both Permissions Hub and App Permissions takes less than half the time it used to.

Bug: 124052170
Test: Open Permissions Hub, change filters, select items and go back
Test: Open App Permissions Screen.
Change-Id: Ifd5d2b157793e6f277e185e0c3055e8f07f3f0d7