Convert usage preferences from nested LinearLayout to ConstraintLayout.
The initial layout of the Permissions Hub has a very slow
measure/layout phase.  I believe this is due to having a large number
of complicated layouts.  The official documentation recommends
replacing LinearLayouts with a layout_weight (which these did) with
ConstraintLayouts.  This commit makes that change.  It should not
change functionality at all.

On my dev device, this seems to speed up the initial layout phase by
15-20%.

Bug: 125332206
Test: Use systrace to time enabling "Show system apps".
Test: Manually compare this to the old screen and they look about the
same to me.

Change-Id: I33849a6a1249822d1e13e204aeff3cc728a32762