Replace SmsDefaultDialog with RequestRoleActivity.
For apps targeting pre-Q, we allow specifying the package name of the
application itself, or another package name if the calling app is the
current default SMS app. Omitting the package name will launch the
settings activity instead. For apps targeting Q or later, we'll cancel
the activity launch in ActivityStackSupervisor.

Bug: 124452117
Bug: 131204827
Test: Open an SMS app and observe the dialog
Change-Id: I8d669164815b9c96e881b790038a47fd65976560