Re-grandfather storage perms
Apprantly some dogfooders got storage permissions that were not
whitelisted. This should have never happened.

Non-whitelisted permission on pre-Q app now means that the pemission is
not grantable anymore, hence the effect of this state is much more
severe than before.

Hence force grandfathering at this time to make sure everybody is back
in a good state.

Fixes: 134508486
Test: Installed P -> Installed Whatsapp -> Granted storage perm -> Upgraded
to Q -> Checked that storage perm is grandfathered.
Change-Id: Iaa667644d6c3ae8db234b96143b67dcd33b773df