Unset even restricted permissions
We do not create AppPermissionGroup objects for restricted permissions
as the user cannot change those.

If a permission is restricted, but needs a review we still need to mark
it as reviewed as otherwise the app won't start. As the user cannot
change it, leaving the permission in the default state without review is
fine.

Test: Installed pre-23 app with restricted permissions and reviewed
      permissions -> app started after review
Fixes: 134500512
Change-Id: If458e739ecf980d2521e87828b75b879aec289c5