Fixed custom permission ui not updating
Fixed issue that caused the AppPermissionsFragment's "additional
permissions" list icon to not move in response to the state of the
permissions inside it.

Fixes: b/140711003

Test: go to the app permissions screen for an app with permissions in
at least one custom permission group. Grant and deny these permissions.
The "Additional permission" line item should reflect the proper number
of permissions, and should show in the allowed, denied, or both
sections, depending on the permission state.

Change-Id: I41590cfe9aff81739e4b87f38512106e7b14f874