Fix crash when role data isn't loaded after restoring state.
In the case where we have a saved state but view model is destroyed,
ListView.getCheckedItemPosition() may return a valid position when we
haven't got the item loaded yet. This change ensures we don't enable
the positive button before role data is loaded, and updates the UI
again when we have the data available.

Fixes: 141367460
Test: presubmit
Change-Id: Ia24cb6396f9038cac4f9bd490705ebca2c4b4a4a