Start special location activities as user
Start the Location Provider and LocationControllerExtraPackage dialogs
and activities as the user of the app selected, in order to open the
correct account in Location History.

Fixes: 143992879

Test: On a device with Google Location History and a work profile,
navigate to the location permissions screen, and select both the work
and personal versions of Google Location History. The personal version
should open the personal account's location history, and the work
version the work account's.

Change-Id: I45b5601b8eee4e9d0c8fb69780f0547a98d13cba