Grant Call audio access to default dialer role
The call audio access will be granted to default dialer apps that require such capability.
This change also added InCallService as a requirement for Dialer role, so we avoid some role abuse.
Also include a string informing that the user will be granting the call audio to the default dialer application.

Test: Compilation and manual test
Bug: 135197853

Change-Id: I77824569091ef964fcff67c663ef370d21841db0