Reduce binder load for UpdateUserSensitive, retry on fail
If getting updating one uid, get the Launch Intent for just that
package. Retry the updating if we catch an exception (specifically
DeadSystemException), after a short delay, up to 3 times.
Also reduces binder calls to LightPackageInfoLiveData. If the UserPackageInfos
livedata is active, then it is much faster for the LightPackageInfos to search
its data for their package, rather than going to the system to retrieve it.

Test: Install app- ensure it is marked as user sensitvie
Fixes: 154796729

Change-Id: I12380ca7f4437017bd0f76f85241b2ba61583b03