Ensure Auto Revoke screen reflects permission grants
Ensure the auto revoked flags are in the flag mask, and are cleared when
granting/revoking permissions
Ensure that when permissions are no longer auto revoked, they are
removed from the auto revoke livedata
Ensure the Auto Revoke livedata creates a full copy of its map, and sets
the value to that, instead of a shallow copy of the values, which are
lists.

Fixes: 153689339
Test: Go to the auto revoked screen, grant a permission which was listed
as auto revoked, and return to the auto revoked screen. The line item
should update to reflect the correct number of revoked permissions

Change-Id: Ia45382e6952e3fc3d7b3446999355fa7dc118cca