Remove open button from auto revoke page, remove used apps
Removes the "open" button from the auto revoke screen, as it is covered
by the app info page, and is not as important as the remove button. Also
removes used apps from the auto revoke screen.

Bug: 153607914
Test: Set threshold to 60 seconds. Get app auto revoked. go to auto
revoke screen, verify it is there. Open app, then navigate back to auto
revoke screen. Verify it is gone. Leave screen, wait 60 seconds, then go
back. Verify it is back.

Change-Id: Id2fbfc3bdf8e4b0046ed41091dcf3292b898761e