Ensure PermStateLiveData registers listener on first load
Ensure that PermStateLiveData checks if it has active observers before
adding a permission listener, not check if registeredUid is null. This
is because registeredUid can be null while the liveData is active (it
happens on first load, when the liveData doesn't know the package UID
yet)

Test: Go to the app info page of an app with denied location permission,
quickly click on the "permissions", then "location", and change from
deny to "ask every time", then go back. The location permission should
move to the "ask every time" category.
Fixes: 159354186

Change-Id: I11244268d537063396370aa07ca1f2efe4c0656e