Fix regression (infinite loop) in TV launch path
If we send ACTION_MANAGE_APP_PERMISSION and don't get a matching form
factor, we're going to reset the intent action and then just try to
launch the same thing again. This is not going to go well and can break
a device until the framework gives up and kills PermissionController.

The solution is to just add a TV case to the same place as the Auto one
that already exists here.

Test: atest android.permission3.cts
Test: atest android.permission2.cts
Test: atest android.permission.cts
Test: atest android.os.cts.AutoRevokeTest
Bug: 155356284
Change-Id: I1830737cdfd933060f800c44b9e599ae7c65d6a1